% Generated by roxygen2 (4.0.2): do not edit by hand
\name{barcodes.to.reads}
\alias{barcodes.to.reads}
\title{Rearrange columns with counts for clonesets.}
\usage{
barcodes.to.reads(.data)
}
\arguments{
\item{.data}{Data frame with columns "Barcode.count" and "Read.count".}
}
\value{
Data frame with new "Read.count" and "Percentage" columns.
}
\description{
Replace Read.count with Barcode.count, recompute Percentage and sort data.
}

