% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{cloneset.stats}
\alias{cloneset.stats}
\alias{repseq.stats}
\title{MiTCR data frame basic statistics.}
\usage{
cloneset.stats(.data, .head = 0)

repseq.stats(.data, .head = 0)
}
\arguments{
\item{.data}{tcR data frames or a list with tcR data frames.}

\item{.head}{How many top clones use to comput summary.}
}
\value{
if \code{.data} is a data frame, than numeric vector with statistics. If \code{.data} is 
a list with data frames, than matrix with statistics for each data frame.
}
\description{
Compute basic statistics of TCR repertoires: number of clones, number of clonotypes, 
number of in-frame and out-of-frame sequences, summary of "Read.count", "Umi.count" and other.
}
\examples{
\dontrun{
# Compute basic statistics of list with data frames.
cloneset.stats(immdata)
repseq.stats(immdata)
}
}
