% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crosses.R
\name{convergence.index}
\alias{convergence.index}
\title{Compute convergence characteristics of repertoires.}
\usage{
convergence.index(.alpha, .beta, .col.nuc = "CDR3.nucleotide.sequence",
  .col.aa = "CDR3.amino.acid.sequence")
}
\arguments{
\item{.alpha}{Either data frame with columns \code{.col.nuc} and \code{.col.aa} or list with such data frames.}

\item{.beta}{Either data frame or none.}

\item{.col.nuc}{Name of the column with nucleotide sequences.}

\item{.col.aa}{Name of the columnw ith aminoacid sequences.}
}
\value{
If \code{.alpha} is data frame, than integer vector of length 2 with . If \code{.alpha} is a list
than matrix M with M[i,j] = convergence.index(.alpha[[i]], .alpha[[j]]).
}
\description{
Get a number of rows with similar aminoacid sequence but different nucleotide sequence.
}

