% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{.add.legend}
\alias{.add.legend}
\title{Internal function. Add legend to a grid of plots and remove legend from all plots of a grid.}
\usage{
.add.legend(.vis.list, .vis.nrow = 2, .legend.ncol = 1)
}
\arguments{
\item{.vis.list}{A list with ggplot2 plots.}

\item{.vis.nrow}{Number of rows of the resulting grid with plots.}

\item{.legend.ncol}{Number of columns in the shared legend.}
}
\description{
Given a list of ggplot2 plots, remove legend from each of them and return 
grid of such plots plus legend from the first vis. Suitable for plots
with similar legends.
}
