% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{.column.choice}
\alias{.column.choice}
\title{Choose the right column.}
\usage{
.column.choice(x, .verbose = T)
}
\arguments{
\item{x}{Character vector with column IDs.}

\item{.verbose}{If T then print the error mesasge.}
}
\value{
Character.
}
\description{
Choose the right column.
}
