% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{.fix.listnames}
\alias{.fix.listnames}
\title{Fix names in lists.}
\usage{
.fix.listnames(.datalist)
}
\arguments{
\item{.datalist}{List with data frames.}
}
\value{
List with fixed names.
}
\description{
Fix names in lists.
}
