% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{.verbose.msg}
\alias{.verbose.msg}
\title{Print the given message if second parameter is a TRUE.}
\usage{
.verbose.msg(.message, .verbose = T)
}
\arguments{
\item{.message}{Character vector standing for a message.}

\item{.verbose}{If T then print the given mesasge.}
}
\value{
Nothing.
}
\description{
Print the given message if second parameter is a TRUE.
}
