% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatools.R
\name{fix.alleles}
\alias{fix.alleles}
\title{Fix alleles - remove allele informatio from columns with genes.}
\usage{
fix.alleles(.data)
}
\arguments{
\item{.data}{Either tcR data frame or a list with tcR data frames.}
}
\description{
Fix alleles - remove allele informatio from columns with genes.
}

