% Generated by roxygen2 (4.0.2): do not edit by hand
\name{gc.content}
\alias{gc.content}
\title{GC-content of a nucleotide sequences.}
\usage{
gc.content(.nucseq)
}
\arguments{
\item{.nucseq}{Character vector of nucletoide sequences.}
}
\value{
Numeric vector of \code{length(.nucseq)}.
}
\description{
Compute the GC-content (proportion of G-C nucleotide in a sequence).
}

