% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strtools.R
\name{get.all.substrings}
\alias{get.all.substrings}
\title{Get all substrings for the given sequence.}
\usage{
get.all.substrings(.seq, .min.len = 3, .table = T)
}
\arguments{
\item{.seq}{Sequence for splitting to substrings.}

\item{.min.len}{Minimal length of output sequences.}

\item{.table}{if T then return data frame with substrings and positions of their ends in the .seq.}
}
\value{
Character vector or data frame with columns "Substring", "Start" and "End".
}
\description{
Get all substrings for the given sequence.
}

