% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph.R
\name{get.people.names}
\alias{get.people.names}
\title{Get people names of the given vertices.}
\usage{
get.people.names(.G, .V = V(.G), .paste = T)
}
\arguments{
\item{.G}{Repertoire graph.}

\item{.V}{Indices of vertices.}

\item{.paste}{If TRUE than concatenate people names to one string, else get a character vector of names.}
}
\value{
Character vector of length .V or list of length .V.
}
\description{
Get people names of the given vertices.
}

