% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmers.R
\name{gibbs.sampler}
\alias{gibbs.sampler}
\title{Gibbs Sampler.}
\usage{
gibbs.sampler(.data, .k = 5, .niter = 500, .meat = T)
}
\arguments{
\item{.data}{Vector of characters or data.frame of characters (1st col) and their numbers (2nd col) if .meat == T.}

\item{.k}{Motif's length.}

\item{.niter}{Number of iterations.}

\item{.meat}{Should function use the count of every sequence or not.}
}
\value{
Vector of possible motifs.
}
\description{
Perform the Gibbs Sampler method for finding frequent motifs in the given vector of strings or data.frame.
Each string splitted to kmers with the given length of motif.
}

