% Generated by roxygen2 (4.0.2): do not edit by hand
\name{has.class}
\alias{has.class}
\title{Check if a given object has a given class.}
\usage{
has.class(.data, .class)
}
\arguments{
\item{.data}{Object.}

\item{.class}{String naming a class.}
}
\value{
Logical.
}
\description{
Check if a given object has a given class.
}

