% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/kmers.R
\name{kmer.profile}
\alias{kmer.profile}
\alias{kmers.profile}
\title{Profile of sequences of equal length.}
\usage{
kmer.profile(.data, .names = rep('Noname', times=length(.data)), .meat = T, .verbose = F)
}
\arguments{
\item{.data}{Either list with elements of one of the allowed classes or object with one of the class:
data.frame with first column with character sequences and second column with number of sequences if .meat = T; character vector,}

\item{.names}{Names for each sequence / row in the \code{.data}.}

\item{.meat}{Use number of sequences (second column of \code{.data} data frame) or not.}

\item{.verbose}{If T than print processing output.}
}
\value{
Return data frame with first column "Symbol" with all possible symbols in the given sequences
and other columns with names "1", "2", ... for each position with percentage for each symbol.
}
\description{
Return profile for the given character vector or a data frame with sequences
of equal length or list with them.
}

