% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/measures.R
\name{loglikelihood}
\alias{loglikelihood}
\title{Log-likelihood.}
\usage{
loglikelihood(.data, .base = 2, .do.norm = NA, .laplace = 1e-12)
}
\arguments{
\item{.data}{Vector for distribution or counts.}

\item{.base}{Logarightm's base for the loglikelihood.}

\item{.do.norm}{Parameter to the \code{check.distribution} function.}

\item{.laplace}{Laplace correction, Parameter to the \code{check.distribution} function.}
}
\value{
Loglikelihood value.
}
\description{
Compute the log-likelihood of the given distribution or vector of counts.
}

