% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatools.R
\name{matrixdiagcopy}
\alias{matrixdiagcopy}
\title{Copy the up-triangle matrix values to low-triangle.}
\usage{
matrixdiagcopy(mat)
}
\arguments{
\item{mat}{Given up-triangle matrix.}
}
\value{
Full matrix.
}
\description{
Copy the up-triangle matrix values to low-triangle.
}

