% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{parse.cloneset}
\alias{parse.cloneset}
\title{Parse input table files with immune receptor repertoire data.}
\usage{
parse.cloneset(.filename, .nuc.seq, .aa.seq, .count, .reads, .barcodes, .vgenes,
  .jgenes, .dgenes, .vend, .jstart, .dalignments, .vd.insertions,
  .dj.insertions, .total.insertions, .skip = 0, .sep = "\\t")
}
\arguments{
\item{.filepath}{Path to the input file with cloneset data.}
}
\value{
Data frame with immune receptor repertoire data. See \link{\code{parse.file}} for more details.
}
\description{
General parser for cloneset table files.
}
\examples{
\dontrun{
# Parse file in "~/mitcr/immdata1.txt" as a MiTCR file.
immdata1 <- parse.file("~/mitcr/immdata1.txt", 'mitcr')
}
}
\seealso{
\link{\code{parse.file}}
}

