% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{parse.file}
\alias{parse.file}
\title{Parse input file with the given filename to a data frame.}
\usage{
parse.file(.filepath, .barcode.flag = F, .i = 1, .all = 1)
}
\arguments{
\item{.filepath}{Path to the input file with TCR data.}

\item{.barcode.flag}{If T than load MiTCR data.frames with new column Barcode.count
and without column Min.quality.}

\item{.i,.all}{Don't supply it, they are just for the output.}
}
\value{
Data.frame with TCR data.
}
\description{
Load the MITCR TCR data from the file with the given filename
to a data frame.
}
\examples{
\dontrun{
# Parse file in "~/data/immdata1.txt".
immdata1 <- parse.file("~/data/immdata1.txt")
}
}
\seealso{
\link{parse.file.list}, \link{parse.folder}
}

