% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{parse.folder}
\alias{parse.file}
\alias{parse.file.list}
\alias{parse.folder}
\alias{parse.migec}
\alias{parse.mitcr}
\alias{parse.mitcrbc}
\title{Parse input table files with immune receptor repertoire data.}
\usage{
parse.folder(.folderpath, .format = c("mitcr", "mitcrbc", "migec"), ...)
}
\arguments{
\item{.folderpath}{Path to the folder with text cloneset files.}

\item{.format}{String specifing input format of files. Parsers for MiTCR output and MiGEC output are available.}

\item{...}{Parameters passed to \code{parse.cloneset}.}

\item{.filepath}{Path to the input file with cloneset data.}

\item{.filenames}{Vector or list with paths to files with cloneset data.}
}
\value{
Data frame with immune receptor repertoire data and following columns:
}
\description{
Load the MITCR TCR data from the file with the given filename
to a data frame. For general parser see \code{\link{parse.cloneset}}.
}
\examples{
\dontrun{
# Parse file in "~/mitcr/immdata1.txt" as a MiTCR file.
immdata1 <- parse.file("~/mitcr/immdata1.txt", 'mitcr')
# Parse files "~/data/immdata1.txt" and "~/data/immdat2.txt" as MiGEC files.
immdata12 <- parse.file.list(c("~/data/immdata1.txt",
                             "~/data/immdata2.txt"), 'migec')
# Parse all files in "~/data/" as MiGEC files.
immdata <- parse.folder("~/data/", 'migec')
}
}
\seealso{
\link{\code{parse.cloneset}}
}

