% Generated by roxygen2 (4.0.2): do not edit by hand
\name{permutedf}
\alias{permutedf}
\alias{unpermutedf}
\title{Shuffling data frames.}
\usage{
permutedf(.data)

unpermutedf(.data)
}
\arguments{
\item{.data}{MiTCR data.frame or list of such data frames.}
}
\value{
Shuffled data.frame or un-shuffled data frame if \code{.data} is a data frame, else list of such data frames.
}
\description{
Shuffle the given data.frame and order it by the Read.count column or un-shuffle
a data frame and return it to the initial order.
}

