% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repdiversity.R
\name{repDiversity}
\alias{repDiversity}
\title{General function for the repertoire diversity estimation.}
\usage{
repDiversity(..., .method = c("chao1", "gini.simp", "inv.simp", "gini", "div",
  "entropy"), .quant = c("read.count", "bc.count", "read.prop", "bc.prop"),
  .q = 5, .do.norm = NA, .laplace = 0)
}
\arguments{
\item{.method}{Which method to use for the diversity estimation. See "Details" for methods.}

\item{.data}{Input clonesets or a list(s) of clonesets.}

\item{.seq}{Which column to use for overlap:}
}
\description{
General interface to all diversity functions.
}
\seealso{
\link{diversity}, \link{entropy}
}

