% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/repoverlap.R
\name{repOverlap}
\alias{repOverlap}
\title{General function for the repertoire overlap evaluation.}
\usage{
repOverlap(..., .method = c("exact", "hamm", "lev", "jaccard", "morisita",
  "cosine", "tversky", "overlap", "horn"), .seq = c("nuc", "aa"),
  .quant = c("read.count", "bc.count", "read.prop", "bc.prop"), .vgene = F,
  .norm = T, .a = 0.5, .b = 0.5, .do.unique = T, .do.norm = T,
  .laplace = 0, .verbose = T)
}
\arguments{
\item{...}{Input clonesets or a list(s) of clonesets.}

\item{.method}{Which method to use for the overlap evaluation. See "Details" for methods.}

\item{.seq}{Which column to use for overlap:}
}
\description{
General interface to all overlap functions.
}
\seealso{
\link{intersect}, \link{similarity}
}

