% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatools.R
\name{repSave}
\alias{repSave}
\title{Save tcR data frames to disk.}
\usage{
repSave(.data, .compress = F, .names = "", .folder = "./")
}
\arguments{
\item{.data}{Either tcR data frame or a list of tcR data frames.}

\item{.compress}{If T than compress text files with GZIP.}

\item{.names}{Names of output files. By default it's an empty string so names will be taken from names of the input list.}

\item{.folder}{Path to the folder with output files.}
}
\description{
Save repertoire files to text or gzipped text files. You can read them by the `parse.tcr` function.
}

