% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/strtools.R
\name{reverse.string}
\alias{reverse.string}
\title{Reverse given character vector by the given n-plets.}
\usage{
reverse.string(.seq, .n = 1)
}
\arguments{
\item{.seq}{Sequences.}

\item{.n}{By which n-plets we should reverse the given strings.}
}
\value{
Reversed strings.
}
\description{
Reverse given character vector by the given n-plets.
}
\examples{
reverse.string('abcde')  # => "edcba"
reverse.string('abcde', 2)  # => "debca"
}

