% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datatools.R
\name{sample.clones}
\alias{sample.clones}
\title{Get a random subset from a data.frame.}
\usage{
sample.clones(.data, .n, .replace = T)
}
\arguments{
\item{.data}{Data.frame or a list with data.frames}

\item{.n}{Sample size if integer. If in bounds [0;1] than percent of rows to extract. "1" is a percent, not one row!}

\item{.replace}{if T then choose with replacement, else without.}
}
\value{
Data.frame of nrow .n or a list with such data.frames.
}
\description{
Sample rows of the given data frame with replacement.
}

