% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample2D}
\alias{sample2D}
\title{Get a sample from matrix with probabilities.}
\usage{
sample2D(.table, .count = 1)
}
\arguments{
\item{.table}{Numeric matrix or data frame with probabilities and columns and rows names.}

\item{.count}{Number of sample to fetch.}
}
\value{
Character matrix with nrow == .count and 2 columns. row[1] in row.names(.table), row[2] in colnames(.table).
}
\description{
Get a sample from matrix or data frame with pair-wise probabilities.
}

