% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/docdata.R
\docType{data}
\name{segments.list}
\alias{genesegments}
\alias{segments.list}
\title{Segment data.}
\format{\code{genesegments} is a list with data frames.}
\description{
\code{segments} is a list with 5 data frames with data of human alpha-beta chain segments.
Elements names as "TRAV", "TRAJ", "TRBV", "TRVJ", "TRVD". Each data frame consists of 5 columns:

- V.allelles / J.allelles / D.allelles - character column with names of V/D/J-segments.

- CDR3.position - position in the full nucleotide segment sequence where CDR3 starts.

- Full.nucleotide.sequence - character column with segment CDR1-2-3 sequence.

- Nucleotide.sequence - character column with segment CDR3 sequences.

- Nucleotide.sequence.P - character column with segment CDR3 sequences with P-insertions.
}
\examples{
\dontrun{
data(genesegments)
genesegments$Nucleotide.sequence[segments$TRBV[,1] == "TRBV10-1"]
}
}

