% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph.R
\name{set.group.vector}
\alias{get.group.names}
\alias{set.group.vector}
\title{Set group attribute for vertices of a mutation network}
\usage{
set.group.vector(.shared.rep, .G, .attr.name, .groups)
}
\arguments{
\item{.shared.rep}{Shared repertiore of sequences.}

\item{.G}{Graph that was created based on \code{.shared.rep}.}

\item{.attr.name}{Name of the new vertex attribute.}

\item{.groups}{List with integer vector with indices of subjects for each group.}
}
\value{
igraph object with new vertex attribute \code{.attr.name} with binary strings.
}
\description{
asdasd
}
\examples{
\dontrun{
data(twb)
twb.shared <- shared.repertoire(twb)
G <- mutation.network(twb.shared)
G <- set.group.vector(twb.shared, G, "twins", list(A = c(1,2), B = c(3,4)))  # <= refactor this
get.group.names(G, "twins", 1)  # "A|B"
get.group.names(G, "twins", 300)  # "A"
# Because we have only two groups, we can assign more readable attribute.
V(G)$twin.names <- get.group.names(G, "twins")
V(G)$twin.names[1]  # "A|B"
V(G)$twin.names[300]  # "A"
}
}

