% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datatools.R
\name{set.pb}
\alias{add.pb}
\alias{set.pb}
\title{Simple functions for manipulating progress bars.}
\usage{
set.pb(.max)

add.pb(.pb, .value = 1)
}
\arguments{
\item{.max}{Length of the progress bar.}

\item{.pb}{Progress bar object.}

\item{.value}{Value to add to the progress bar.}
}
\value{
Progress bar (for set.pb) or length-one numeric vector giving the previous value (for add.pb).
}
\description{
Set the progress bar with the given length (from zero) or add value to the created progress bar.
}

