% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graph.R
\name{set.people.vector}
\alias{get.people.names}
\alias{set.people.vector}
\title{Set and get attributes of repertoire graph related to source people.}
\usage{
set.people.vector(.G, .shared.rep)
}
\arguments{
\item{.G}{Repertoire graph.}

\item{.shared.rep}{Shared repertoire.}
}
\value{
New graph with 'people' and 'npeople' vertex attributes.
}
\description{
Set vertice attributes 'people' and 'npeople' for every vertex in the given graph.
Attribute 'people' is a binary string indicating in which repertoire sequence are
found. Attribute 'npeople' is a integer indicating number of repertoires, in which
this sequence has been found.
}

