% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataproc.R
\name{set.rank}
\alias{set.index}
\alias{set.rank}
\title{Set new columns "Rank" and "Index".}
\usage{
set.rank(.data, .col = "Read.count")
}
\arguments{
\item{.data}{Data frame or list with data frames.}

\item{.col}{Character vector with name of the column to use for ranking or indexing.}
}
\value{
Data frame with new column "Rank" or "Index" or list with such data frames.
}
\description{
Set new columns "Rank" and "Index":

set.rank <==> .data$Rank = rank(.data[, .col], ties.method = 'average')

set.index <==> .data$Index = 1:nrow(.data) in a sorted data frame by \code{.col}
}

