% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docdata.R
\docType{data}
\name{twinsdata}
\alias{twinsdata}
\alias{twa}
\alias{twb}
\title{Twins alpha-beta chain data}
\format{
\code{twa} and \code{twb} are lists of 4 data frames with 10000 row in each.
}
\description{
\code{twa.rda}, \code{twb.rda} - data frames with downsampled to the 10000 most 
abundant clonesets and 4 samples data of twins data (alpha and beta chains).
Link: http://labcfg.ibch.ru/tcr.html
}
\examples{
\dontrun{
data(twa)
data(twb)
}
}
