% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{vis.clonal.space}
\alias{vis.clonal.space}
\title{Visualise occupied by clones homeostatic space among subjects or groups.}
\usage{
vis.clonal.space(.clonal.space.data, .groups = NULL)
}
\arguments{
\item{.clonal.space.data}{from the \code{find.clonotypes} function, i.e. data frame with first
columns with sequences (nucleotide or amino acid) and other columns are columns with frequency / count
for each time point for each clone.}

\item{.groups}{}
}
\value{
ggplot object.
}
\description{
Visualise clonal dynamics (i.e., changes in frequency or count) with error bars of given
clones among time points.
}
\seealso{
\link{clonal.space.homeostasis}
}

