% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{vis.count.len}
\alias{vis.count.len}
\title{Plot a histogram of lengths.}
\usage{
vis.count.len(.data, .ncol = 3, .name = "", .col = "Read.count")
}
\arguments{
\item{.data}{Data frame with columns 'CDR3.nucleotide.sequence' and 'Read.count' or list with such data frames.}

\item{.ncol}{If .data is a list, than number of columns in a grid of histograms for each data frame in \code{.data}. Else not used.}

\item{.name}{Title for this plot.}

\item{.col}{Name of the column to use in computing the lengths distribution.}
}
\value{
ggplot object.
}
\description{
Plot a histogram of distribution of lengths of CDR3 nucleotide sequences. On y-axis are sum of read counts for each length.
}
\details{
If \code{.data} is a data frame, than one histogram will be plotted. Is \code{.data} is a list, than grid of histograms
will be plotted.
}
\examples{
\dontrun{
load('immdata.rda')
# Plot one histogram with main title.
vis.count.len(immdata[[1]], 'Main title here')
# Plot a grid of histograms with 2 columns.
vis.count.len(immdata, 2)
}
}
