% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{vis.pca}
\alias{vis.pca}
\title{PCA result visualisation}
\usage{
vis.pca(.data, .groups = NA)
}
\arguments{
\item{.data}{Result from prcomp() function or a data frame with two columns 'First' and 'Second'
stands for the first PC and the second PC.}

\item{.groups}{List with names for groups and indices of the group members. If NA than each
member is in the individual group.}
}
\value{
ggplot object.
}
\description{
Plot the given pca results with colour divided by the given groups.
}

