% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plots.R
\name{vis.top.proportions}
\alias{vis.top.proportions}
\title{Visualisation of top clones proportions.}
\usage{
vis.top.proportions(.data, .head = c(10, 100, 1000, 10000, 30000, 1e+05,
  3e+05, 1e+06), .col = "Read.count")
}
\arguments{
\item{.data}{Data frame with clones.}

\item{.head}{Integer vector of clones for the \code{.head} parameter for the \code{top.proportion} function.}

\item{.col}{Parameter \code{.col} for the \code{top.proportion} function.}
}
\description{
Visualisation of proportion of the top clones.
}
\examples{
\dontrun{
vis.top.proportions(immdata)
}
}
\seealso{
\code{top.proportion}
}

