% Generated by roxygen2 (4.0.2): do not edit by hand
\name{assymetry}
\alias{assymetry}
\title{Normalised log assymetry.}
\usage{
assymetry(.alpha, .beta = NULL, .by = "CDR3.nucleotide.sequence")
}
\arguments{
\item{.alpha}{First mitcr data.frame or a list with data.frames.}

\item{.beta}{Second mitcr data.frame or NULL if \code{.alpha} is a list.}

\item{.by}{Which column use to merge. See "Details".}
}
\value{
Value of the normalised log assymetry measure for the given data.frames.
}
\description{
Compute the value of the normalised log assymetry measure for the given data.frames
of the counts of shared clones.
}
\details{
Merge two data frames by the given column and compute
value Sum(Log((Percentage for shared clone S from alpha) / (Percentage for shared clone S from beta))) / (# of shared clones).
}

