% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.R
\name{barcodes.to.reads}
\alias{barcodes.to.reads}
\title{Rearrange columns with counts for clonesets.}
\usage{
barcodes.to.reads(.data)
}
\arguments{
\item{.data}{Data frame with columns "Umi.count" and "Read.count".}
}
\value{
Data frame with new "Read.count" and "Percentage" columns.
}
\description{
Replace Read.count with Umi.count, recompute Percentage and sort data.
}

