% Generated by roxygen2 (4.0.2): do not edit by hand
\name{check.distribution}
\alias{check.distribution}
\title{Check for adequaty of distrubution.}
\usage{
check.distribution(.data, .do.norm = NA, .laplace = 1, .na.val = 0)
}
\arguments{
\item{.data}{Numeric vector of values.}

\item{.do.norm}{One of the three values - NA, T or F. If NA than check for distrubution (sum(.data) == 1)
and normalise if needed with the given laplace correction value. If T than do normalisation and laplace
correction. If F than don't do normalisaton and laplace correction.}

\item{.laplace}{Value for laplace correction.}

\item{.na.val}{Replace all NAs with this value.}
}
\value{
Numeric vector.
}
\description{
Check if the given .data is a distribution and normalise it if necessary with optional laplace correction.
}

