% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatools.R
\name{fix.alleles}
\alias{fix.alleles}
\alias{fix.genes}
\title{Fix alleles / genes by removing allele information / unnecessary colons.}
\usage{
fix.alleles(.data)
}
\arguments{
\item{.data}{tcR data frame.}
}
\description{
Fix alleles / genes by removing allele information / unnecessary colons.
}
