% Generated by roxygen2 (4.0.2): do not edit by hand
\name{matrixdiagcopy}
\alias{matrixdiagcopy}
\title{Copy the up-triangle matrix values to low-triangle.}
\usage{
matrixdiagcopy(mat)
}
\arguments{
\item{mat}{Given up-triangle matrix.}
}
\value{
Full matrix.
}
\description{
Copy the up-triangle matrix values to low-triangle.
}

