% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/input.R
\name{parse.folder}
\alias{parse.folder}
\title{Parse all files to dataframes from the given path to folder.}
\usage{
parse.folder(.folderpath, .barcode.flag = F)
}
\arguments{
\item{.folderpath}{Path to the folder with files with TCR data.}

\item{.barcode.flag}{If T than load MiTCR data.frames with new column Barcode.count
and without column Min.quality.}
}
\value{
List with data frame for the each file in the given folder.
}
\description{
Given the path to a folder with files with TCR data, parse all of them
to a list and return it. List items have names similar to names in the
given vector if filenames.
}
\examples{
\dontrun{
# Parse all files in "~/data/".
immdata <- parse.folder("~/data/")
}
}
\seealso{
\link{parse.file}, \link{parse.file.list}
}

