% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{repSave}
\alias{repSave}
\title{Save tcR data frames to disk as text files or gzipped text files.}
\usage{
repSave(.data, .format = c("txt", "gz"), .names = "", .folder = "./")
}
\arguments{
\item{.data}{Either tcR data frame or a list of tcR data frames.}

\item{.format}{"txt" for simple tab-delimited text tables, "gz" for compressed (gzipped) tables.}

\item{.names}{Names of output files. By default it's an empty string so names will be taken from names of the input list.}

\item{.folder}{Path to the folder with output files.}
}
\description{
Save repertoire files to either text files or gzipped text files.
You can read them later by \code{repLoad} function with \code{.format = "tcr"}.
}
\seealso{
\link{repLoad}
}
