% Generated by roxygen2 (4.0.2): do not edit by hand
\name{revcomp}
\alias{bunch.translate}
\alias{revcomp}
\title{DNA reverse complementing and translation.}
\usage{
revcomp(.seq)

bunch.translate(.seq, .two.way = T)
}
\arguments{
\item{.seq}{Vector of nucleotide sequences.}

\item{.two.way}{If T than translate sequences from both ends (output differes for
out-of-frame sequences).}
}
\value{
Vector of corresponding revese complemented or aminoacid sequences.
}
\description{
Functions for DNA reverse complementing and translation.
}

