% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/docdata.R
\docType{data}
\name{segments.alphabets}
\alias{J_ALPHA_ALPHABET}
\alias{J_BETA_ALPHABET}
\alias{V_ALPHA_ALPHABET}
\alias{V_BETA_ALPHABET}
\alias{segments.alphabets}
\title{Alphabets of V-J segments.}
\format{Each \code{*_*_ALPHABET} is a character vector.}
\description{
Character vector with names for segments. With \code{tcR} we provided
two dataset of segments for humans ("./data/human.alphabets.rda") and mouses ("./data/mouse.alphabets.rda").
}
\examples{
\dontrun{
data(human.alphabets)
V_BETA_ALPHABET[1]  #  => "TRBV10-1"
}
}

