% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vis.clonal.dynamics}
\alias{vis.clonal.dynamics}
\title{Visualise clonal dynamics among time points.}
\usage{
vis.clonal.dynamics(.changed, .lower, .upper, .log = T)
}
\arguments{
\item{.changed}{Result from the \code{find.clonotypes} function, i.e. data frame with first
columns with sequences (nucleotide or amino acid) and other columns are columns with frequency / count
for each time point for each clone.}

\item{.lower}{Similar to .changed but values are lower bound for clonal count / frequency.}

\item{.upper}{Similar to .changed but values are upper bound for clonal count / frequency.}

\item{.log}{If T than log-scale y-axis.}
}
\value{
ggplot object.
}
\description{
Visualise clonal dynamics (i.e., changes in frequency or count) with error bars of given
clones among time points.
}

