% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plots.R
\name{vis.clonal.space}
\alias{vis.clonal.space}
\title{Visualise occupied by clones homeostatic space among subjects or groups.}
\usage{
vis.clonal.space(.clonal.space.data, .groups = NULL)
}
\arguments{
\item{.clonal.space.data}{Data from the \code{fclonal.space.homeostasis} function.}

\item{.groups}{List of named character vector with names of subjects
in \code{.clonal.space.data} for grouping them together.}
}
\value{
ggplot object.
}
\description{
Visualise which clones how much space occupy.
}
\seealso{
\link{clonal.space.homeostasis}
}

